Add parallel support.
Virtual Machines can now be brought up in parallel.
Driver encapsulates the logic to run commands on remote sites in a thread
safe manner. There is a driver per (gateway-site-username), so you can
theorically starts VMs on different platform in the same run.

In the future driver could be abstracted. For the moment it is tight to
net/ssh.