Add support for bridged network
G5K provider now supports two types of VM networking :

* nat. VM can access the outside through the native NAT capability of kvm
* bridge. VM are given an IP from a G5K subnet. They can commuicate with each
other using their assigned IP.

Under the hood, one subnet is reserved if at least one VM requires a bridge
network. More precisely the first VM to boot will also reserve the subnet. At
any time, the number of VMs using the subnet is tracked so that when no more VM
is using it, the subnet job is deleted. This way the subnet management is
transparent to the user.

M